move utils tests to proper module
